Add If/else
Plan and add a feature to one of the web pages in your personal site that incorporates content or functionality created by a series of if, if/else, and/or else if statements, or by a switch statement. If your page requires an event listener, create one that's backward compatible with older versions of Internet Explorer. View and test your page in one or more browsers as appropriate to ensure it works as you expect.