Add a Math page to the website
Expand your individual website to include a page that calculates the time elapsed since a date entered by a user. The page should include a form that allows users to enter a day,month, and year. The page should then calculate and display the elapsed time in years, months, and days. Note that your program must include code to convert day values in excess of 31 into months, and months in excess of 12 into years.