Add Security Changes using RegEx
Enhance the personal website you've created in the preceding chapters of this book to prevent security issues. To do this, review each form field in your site and identify any additional validation that you could reasonably add. In particular, look for situations where you could use a regular expression to limit the allowable characters in order to exclude characters used in creating JavaScript code. Use word-processing software to create a table like Table 9-4 (on page 679), which includes an example. Add a row for each form field in your website, and then save this document in the root of your site. It doesn't have to be linked to from your main site structure, but does need to be present in your submission for full credit on this assignment. Points will be deducted for failure to include this planning document.